refactor: :recycle: refactor profile info
